Client: moves Actions & Reducers back to src/ & updates paths to them
