+++ client/src/store/store.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import rootReducer from './Reducers';
+import rootReducer from '../Reducers';
 
 const store = createStore(
   rootReducer,
